move mail to internal
