+++ internal/mail/gmail/gmailauth.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019 Storj Labs, Inc.
 // See LICENSE for copying information
 
-package mail
+package gmail
 
 import (
 	"bytes"
